# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_chienchuanw_asset-manager_internal_models.AllocationByAsset:
    properties:
      asset_type:
        allOf:
        - $ref: '#/definitions/github_com_chienchuanw_asset-manager_internal_models.AssetType'
        description: 資產類型
      market_value:
        description: 市值
        type: number
      name:
        description: 資產名稱
        type: string
      percentage:
        description: 佔比（%）
        type: number
      quantity:
        description: 持有數量
        type: number
      symbol:
        description: 資產代碼
        type: string
    type: object
  github_com_chienchuanw_asset-manager_internal_models.AllocationByType:
    properties:
      asset_type:
        allOf:
        - $ref: '#/definitions/github_com_chienchuanw_asset-manager_internal_models.AssetType'
        description: 資產類型
      count:
        description: 持倉數量
        type: integer
      market_value:
        description: 市值
        type: number
      name:
        description: 資產類型名稱
        type: string
      percentage:
        description: 佔比（%）
        type: number
    type: object
...
  github_com_chienchuanw_asset-manager_internal_models.AllocationSummary:
    properties:
      as_of_date:
        description: 資料日期
        type: string
      by_asset:
        description: 按個別資產分類
        items:
          $ref: '#/definitions/github_com_chienchuanw_asset-manager_internal_models.AllocationByAsset'
        type: array
      by_type:
        description: 按資產類型分類
        items:
          $ref: '#/definitions/github_com_chienchuanw_asset-manager_internal_models.AllocationByType'
        type: array
      currency:
        description: 幣別
        type: string
      total_market_value:
        description: 總市值
        type: number
    type: object
...
  github_com_chienchuanw_asset-manager_internal_models.AssetType:
    enum:
    - cash
    - tw-stock
    - us-stock
    - crypto
    type: string
    x-enum-varnames:
    - AssetTypeCash
    - AssetTypeTWStock
    - AssetTypeUSStock
    - AssetTypeCrypto
  github_com_chienchuanw_asset-manager_internal_models.AssetTypeDeviation:
    properties:
      asset_type:
        description: 資產類型
        type: string
      current_percent:
        description: 當前配置百分比
        type: number
      current_value:
        description: 當前市值
        type: number
      deviation:
        description: 偏離百分比（當前 - 目標）
        type: number
      deviation_abs:
        description: 絕對偏離百分比
        type: number
      exceeds_threshold:
        description: 是否超過閾值
        type: boolean
      target_percent:
        description: 目標配置百分比
        type: number
      target_value:
        description: 目標市值
        type: number
    type: object
...
  github_com_chienchuanw_asset-manager_internal_models.RebalanceCheck:
    properties:
      current_total:
        description: 當前總資產
        type: number
      deviations:
        description: 各資產類型的偏離情況
        items:
          $ref: '#/definitions/github_com_chienchuanw_asset-manager_internal_models.AssetTypeDeviation'
        type: array
      needs_rebalance:
        description: 是否需要再平衡
        type: boolean
      suggestions:
        description: 再平衡建議
        items:
          $ref: '#/definitions/github_com_chienchuanw_asset-manager_internal_models.RebalanceSuggestion'
        type: array
      threshold:
        description: 閾值百分比
        type: number
    type: object
  github_com_chienchuanw_asset-manager_internal_models.RebalanceSuggestion:
    properties:
      action:
        description: 動作：buy（買入）或 sell（賣出）
        type: string
      amount:
        description: 金額（TWD）
        type: number
      asset_type:
        description: 資產類型
        type: string
      reason:
        description: 原因說明
        type: string
    type: object
...
  github_com_chienchuanw_asset-manager_internal_models.UnrealizedPerformance:
    properties:
      asset_type:
        allOf:
        - $ref: '#/definitions/github_com_chienchuanw_asset-manager_internal_models.AssetType'
        description: 資產類型
      cost:
        description: 成本
        type: number
      holding_count:
        description: 持倉數量
        type: integer
      market_value:
        description: 市值
        type: number
      name:
        description: 資產類型名稱
        type: string
      unrealized_pct:
        description: 未實現報酬率
        type: number
      unrealized_pl:
        description: 未實現損益
        type: number
    type: object
  github_com_chienchuanw_asset-manager_internal_models.UnrealizedSummary:
    properties:
      currency:
        description: 幣別
        type: string
      holding_count:
        description: 持倉數量
        type: integer
      total_cost:
        description: 總成本
        type: number
      total_market_value:
        description: 總市值
        type: number
      total_unrealized_pct:
        description: 總未實現報酬率
        type: number
      total_unrealized_pl:
        description: 總未實現損益
        type: number
    type: object
  github_com_chienchuanw_asset-manager_internal_models.UnrealizedTopAsset:
    properties:
      asset_type:
        allOf:
        - $ref: '#/definitions/github_com_chienchuanw_asset-manager_internal_models.AssetType'
        description: 資產類型
      avg_cost:
        description: 平均成本
        type: number
      cost:
        description: 總成本
        type: number
      current_price:
        description: 當前價格
        type: number
      market_value:
        description: 市值
        type: number
      name:
        description: 標的名稱
        type: string
      quantity:
        description: 持有數量
        type: number
      symbol:
        description: 標的代碼
        type: string
      unrealized_pct:
        description: 未實現報酬率
        type: number
      unrealized_pl:
        description: 未實現損益
        type: number
    type: object
...
  internal_api.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  internal_api.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/internal_api.APIError'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Personal finance system API for tracking investment portfolios, cash
    flows, subscriptions, and financial analytics.
  title: Asset Manager API
  version: "1.0"
paths:
  /api/allocation/by-asset:
    get:
      consumes:
      - application/json
      description: 取得按個別資產分類的資產配置
      parameters:
      - default: 20
        description: 回傳數量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_chienchuanw_asset-manager_internal_models.AllocationByAsset'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      summary: 取得按個別資產的配置
      tags:
      - allocation
  /api/allocation/by-type:
    get:
      consumes:
      - application/json
      description: 取得按資產類型分類的資產配置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_chienchuanw_asset-manager_internal_models.AllocationByType'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      summary: 取得按資產類型的配置
      tags:
      - allocation
  /api/allocation/current:
    get:
      consumes:
      - application/json
      description: 取得當前所有持倉的資產配置摘要，包含按資產類型和個別資產的分類
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_chienchuanw_asset-manager_internal_models.AllocationSummary'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      summary: 取得當前資產配置摘要
      tags:
      - allocation
...
  /api/analytics/unrealized/performance:
    get:
      consumes:
      - application/json
      description: 取得當前各資產類型的未實現損益績效
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_chienchuanw_asset-manager_internal_models.UnrealizedPerformance'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      summary: 取得各資產類型未實現績效
      tags:
      - unrealized-analytics
  /api/analytics/unrealized/summary:
    get:
      consumes:
      - application/json
      description: 取得當前所有持倉的未實現損益摘要
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_chienchuanw_asset-manager_internal_models.UnrealizedSummary'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      summary: 取得未實現損益摘要
      tags:
      - unrealized-analytics
  /api/analytics/unrealized/top-assets:
    get:
      consumes:
      - application/json
      description: 取得未實現損益最佳的資產（按未實現損益排序）
      parameters:
      - default: 10
        description: 回傳數量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_chienchuanw_asset-manager_internal_models.UnrealizedTopAsset'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      summary: 取得 Top 未實現損益資產
      tags:
      - unrealized-analytics
...
  /api/rebalance/check:
    get:
      consumes:
      - application/json
      description: 檢查當前資產配置是否偏離目標配置，並提供再平衡建議
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_chienchuanw_asset-manager_internal_models.RebalanceCheck'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.APIResponse'
      summary: 檢查是否需要再平衡
      tags:
      - rebalance
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"